+++ src/main/java/kamitesque/common/blocks/BlockRootCrystal.java
@@ -2,12 +2,14 @@ package kamitesque.common.blocks;
 
 import kamitesque.common.templates.BlockKTTile;
 import kamitesque.common.tiles.TileRootCrystal;
+import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.properties.PropertyBool;
 import net.minecraft.block.state.BlockFaceShape;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumFacing;
@@ -17,32 +19,56 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import thaumcraft.common.blocks.IBlockFacing;
 import thaumcraft.common.lib.SoundsTC;
 
 @SuppressWarnings("deprecation")
-public class BlockRootCrystal extends BlockKTTile {
+public class BlockRootCrystal extends BlockKTTile implements IBlockFacing {
 
     public static String id = "root_crystal";
 
-    public static PropertyBool GROWN = PropertyBool.create("grown");
+    public static PropertyBool IS_GROWING = PropertyBool.create("growing");
 
-    AxisAlignedBB box =  new AxisAlignedBB(
-    0.25, 0.0, 0.25,
+    public static final AxisAlignedBB BOX_UP = new AxisAlignedBB(
+            0.25, 0.0, 0.25,
             0.75, 1.0, 0.75
+    );
+
+    public static final AxisAlignedBB BOX_DOWN = new AxisAlignedBB(
+            0.25, 0.0, 0.25,
+            0.75, 1.0, 0.75
+    );
+
+    public static final AxisAlignedBB BOX_NORTH = new AxisAlignedBB(
+            0.25, 0.25, 0.0,
+            0.75, 0.75, 1.0
+    );
+
+    public static final AxisAlignedBB BOX_SOUTH = new AxisAlignedBB(
+            0.25, 0.25, 0.0,
+            0.75, 0.75, 1.0
+    );
+
+    public static final AxisAlignedBB BOX_WEST = new AxisAlignedBB(
+            0.0, 0.25, 0.25,
+            1.0, 0.75, 0.75
+    );
+
+    public static final AxisAlignedBB BOX_EAST = new AxisAlignedBB(
+            0.0, 0.25, 0.25,
+            1.0, 0.75, 0.75
     );
 
     public BlockRootCrystal() {
         super(Material.GLASS, TileRootCrystal.class, id);
         this.setHardness(0.25F);
         this.setSoundType(SoundsTC.CRYSTAL);
+
+        this.setDefaultState(this.getDefaultState().withProperty(IS_GROWING, false));
     }
 
     public SoundType getSoundType() {
         return SoundsTC.CRYSTAL;
-    }
-
-    public boolean canPlaceBlockAt(World worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos.down()).getBlock().equals(Blocks.BEDROCK);
     }
 
     public boolean isFullCube(IBlockState state) {
@@ -55,7 +81,7 @@ public class BlockRootCrystal extends BlockKTTile {
 
     @Override
     protected BlockStateContainer createBlockState() {
-        return new BlockStateContainer(this, GROWN);
+        return new BlockStateContainer(this, IS_GROWING, IBlockFacing.FACING);
     }
 
     @SideOnly(Side.CLIENT)
@@ -69,28 +95,80 @@ public class BlockRootCrystal extends BlockKTTile {
 
     @Override
     public IBlockState getStateFromMeta(int meta) {
+        EnumFacing facing = EnumFacing.byIndex(meta & 7);
+        boolean active = (meta & 8) != 0;
+
         return getDefaultState()
-                .withProperty(GROWN, meta != 0);
+                .withProperty(FACING, facing)
+                .withProperty(IS_GROWING, active);
     }
 
     @Override
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(GROWN) ? 1 : 0;
+        int meta = state.getValue(FACING).getIndex();
+
+        if (state.getValue(IS_GROWING)) {
+            meta |= 8;
+        }
+
+        return meta;
     }
 
-    public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
-        return this.box;
+    @Override
+    public AxisAlignedBB getBoundingBox(IBlockState state,
+                                        IBlockAccess source,
+                                        BlockPos pos) {
+
+        switch (state.getValue(FACING)) {
+
+            case DOWN:
+                return BOX_DOWN;
+
+            case NORTH:
+                return BOX_NORTH;
+
+            case SOUTH:
+                return BOX_SOUTH;
+
+            case WEST:
+                return BOX_WEST;
+
+            case EAST:
+                return BOX_EAST;
+
+            default:
+                return BOX_UP;
+        }
     }
 
-//    public void addCollisionBoxToList(IBlockState state, World worldIn, BlockPos pos, AxisAlignedBB entityBox, List<AxisAlignedBB> collidingBoxes, @Nullable Entity entityIn, boolean isActualState) {
-//
-//        addCollisionBoxToList(pos, entityBox, collidingBoxes, this.box);
-//    }
-//
-//    @Nullable
-//    public AxisAlignedBB getCollisionBoundingBox(IBlockState blockState, IBlockAccess worldIn, BlockPos pos) {
-//        blockState = this.getActualState(blockState, worldIn, pos);
-//        return CLIP_AABB_BY_INDEX[getAABBIndex(blockState)];
-//    }
+    public IBlockState getStateForPlacement(World worldIn, BlockPos pos, EnumFacing facing, float hitX, float hitY, float hitZ, int meta, EntityLivingBase placer)
+    {
+        boolean growing = worldIn.getBlockState(
+                pos.offset(facing.getOpposite())
+        ).getBlock() == Blocks.BEDROCK;
+
+        return this.getDefaultState()
+                .withProperty(FACING, facing)
+                .withProperty(IS_GROWING, growing);
+    }
+
+    public void neighborChanged(IBlockState state,
+                                World world,
+                                BlockPos pos,
+                                Block blockIn,
+                                BlockPos fromPos) {
+
+        EnumFacing facing = state.getValue(FACING);
+
+        BlockPos supportPos = pos.offset(facing.getOpposite());
+
+        if (!world.getBlockState(supportPos).isSideSolid(
+                world,
+                supportPos,
+                facing)) {
+
+            world.destroyBlock(pos, true);
+        }
+    }
 
 }