+++ src/main/java/kamitesque/common/items/ItemAwakenedHoe.java
@@ -1,0 +1,9 @@
+package kamitesque.common.items;
+
+import mod.emt.kami.items.tools.ItemIchoriumPickaxe;
+
+public class ItemAwakenedHoe extends ItemIchoriumPickaxe {
+
+    public static final String id = "awakened_ichorium_hoe";
+
+}