+++ src/main/java/kamitesque/root/Main.java
@@ -1,6 +1,7 @@
 package kamitesque.root;
 
 import kamitesque.init.KTEntities;
+import kamitesque.init.KTRecipes;
 import kamitesque.init.KTResearch;
 import kamitesque.init.KTTiles;
 import kamitesque.network.packets.KTNetwork;
@@ -29,7 +30,7 @@ import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 public class Main {
     public static final String MODID = "kamitesque";
     public static final String NAME = "Kamitesque";
-    public static final String VERSION = "0.0.5-ALPHA";
+    public static final String VERSION = "0.0.7-ALPHA";
 
     public static SimpleNetworkWrapper packetHandler;
 
@@ -59,6 +60,8 @@ public class Main {
 
     @EventHandler
     public void postInit(FMLPostInitializationEvent event) {
+
+        KTRecipes.initFurnace();
 
     }
 }