+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # Kamitesque Changelog
+## 0.0.7-ALPHA
+
+---
+### Added:
+- Root Crystal FX on process completion
+- Root Crystal texture animation
+### Fixed:
+- Mod items not displaying in the creative search tab
+- Infernal furnace crystal cluster smelting bonus not being registered
+- Root Crystal Seed recipe not displaying in the research entry
+- Root Crystal accelerating its growth speed on each new added to the world
+- Ichorium Needle max stack limit != 1
+### Changed:
+- Root Crystal rewrite, can be placed in any direction on any block now
+- Ichorium Needle texture
+---
 ## 0.0.6-ALPHA
 
 ---